Set up visual studio solution for the new project structure. Also cleaned up all the stuff that visual c++ gives warnings about. For some reason GLEW isn't working on this laptop now so I can't test anything.

+++ include/Lexer.h
@@ -41,8 +41,8 @@ namespace grumpy
 		SyntaxParser *parser;
         std::vector<LexToken> lTokens;
         std::vector<Token*> tokensProduced;
-        int currentCharacterIndex;
-        int currentTokenIndex;
+        size_t currentCharacterIndex;
+		size_t currentTokenIndex;
         glm::vec4 currentTokenStartPos;
 
         float skipVelocity;
